+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Catch confirmation form up to recent changes in the long form. Fixes PR-1688.
 * Update the "request sent" page to reflect appropriate metadata for copies as well as loans. Fixes PR-1687.
 * Set copyright type by ID (which works) rather than by value (which does not). Fixes PR-1690.
+* "Request sent" page shows copyright type only for copies, not loans. Fixes PR-1698.
 
 ## [1.6.0](https://github.com/openlibraryenvironment/listener-openurl/tree/v1.6.0) (2024-03-04)